requirements.txt
openai
anthropic
pyyaml
requests
streamlit